+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Changelog for NeoFS Node
 - No redistribution of objects when adding a new `REP 3` node (#3873)
 - Split of a split is allowed (#3867)
 - SN no longer fails to read valid objects with non-payload fields larger than 16K (#3887)
+- Session v2 origin signature verification (#3909)
 
 ### Changed
 - SN returns unsigned responses to requests with API >= `v2.22` (#3785)